Fix error in the Makefile
Dash doesn’t support the `-e` option of the `echo` built-in command.
`echo -e` prints `-e` and a line feed:

```sh
$ echo -e
-e
```

I was unable to compile the project because the generated file
`targets.mk` contained `-e` at the beginning of some lines.
It leads to a Make syntax error.

Dash is the defaut shell in Debian. It's important to support it.